Move fetch from redux functionality
